+++ Tools/west_environment.f90
@@ -1,5 +1,5 @@
 !
-! Copyright (C) 2015-2017 M. Govoni 
+! Copyright (C) 2015-2019 M. Govoni 
 ! This file is distributed under the terms of the
 ! GNU General Public License. See the file `License'
 ! in the root directory of the present distribution,
@@ -37,6 +37,8 @@ CONTAINS
     USE westcom,         ONLY : savedir, logfile, outdir, west_prefix 
     USE base64_module,   ONLY : base64_init 
     USE json_string_utilities, ONLY : lowercase_string
+    USE west_version,    ONLY : start_forpy
+    !USE logfile_mod,     ONLY : clear_log
     !
     CHARACTER(LEN=*), INTENT(IN) :: code
     !
@@ -52,15 +54,16 @@ CONTAINS
 #if defined(__INTEL_COMPILER)
     CALL remove_stack_limit ( )
 #endif
+    CALL start_forpy() 
     !
     ! Input from (-i), output from (-o)
     !
     CALL parse_command_arguments()
-    CALL fetch_input(1,(/1/),.FALSE.)
+    CALL fetch_input_yml(1,(/1/),.FALSE.,.FALSE.)
     !
-    savedir = TRIM(outdir) // trim(west_prefix) // "." // TRIM(lowercase_string(code)) // ".save/"
-    CALL my_mkdir( savedir )
-    logfile = TRIM(savedir) // TRIM(lowercase_string(code))//".json"
+    savedir = TRIM(ADJUSTL(outdir)) // TRIM(ADJUSTL(west_prefix)) // "." // TRIM(lowercase_string(code)) // ".save/"
+    logfile = TRIM(ADJUSTL(savedir)) // TRIM(lowercase_string(code))//".json"
+    CALL my_mkdir( TRIM(ADJUSTL(savedir)) ) 
     !
     ! ... use ".FALSE." to disable all clocks except the total cpu time clock
     ! ... use ".TRUE."  to enable clocks
@@ -113,8 +116,9 @@ CONTAINS
     END IF
     !
     ! Initialize base64 tables  
-    CALL base64_init() 
+    CALL base64_init()
     !
+    !CALL clear_log()
     CALL west_opening_message( code )
 #if defined(__MPI)
     CALL report_parallel_status ( )
@@ -131,6 +135,7 @@ CONTAINS
     USE mp_world,        ONLY : mpime,root,world_comm
     USE mp,              ONLY : mp_barrier
     USE westcom,         ONLY : logfile 
+    USE west_version,    ONLY : end_forpy
     !
     IMPLICIT NONE 
     !
@@ -181,6 +186,8 @@ CONTAINS
       !
     ENDIF
     !
+    CALL end_forpy()
+    !
     CALL mp_barrier(world_comm) 
     !
   END SUBROUTINE
@@ -192,10 +199,12 @@ CONTAINS
     USE json_module,     ONLY : json_file
     USE io_global,       ONLY : stdout
     USE global_version,  ONLY : version_number, svn_revision
-    USE west_version,    ONLY : west_version_number, west_svn_revision
+    USE west_version,    ONLY : west_version_number, west_git_revision
     USE mp_world,        ONLY : mpime,root 
     USE westcom,         ONLY : logfile
     USE base64_module,   ONLY : islittleendian  
+    USE forpy_mod,        ONLY : dict, dict_create 
+    !USE logfile_mod,      ONLY : append_log, itoa, ltoa, dtoa
     !
     ! I/O
     !
@@ -207,11 +216,15 @@ CONTAINS
     INTEGER :: iunit
     CHARACTER(LEN=9)  :: cdate, ctime
     !
+    INTEGER :: IERR
+    CHARACTER(LEN=:),ALLOCATABLE :: s
+    TYPE(dict) :: attr
+    !
     CALL date_and_tim( cdate, ctime )
     !
-    IF ( TRIM (west_svn_revision) /= "unknown" ) THEN
-       WRITE( stdout, '(/5X,"Program ",A," v. ",A," svn rev. ",A," starts on ",A9," at ",A9)' ) &
-       & TRIM(code), TRIM(west_version_number), TRIM (west_svn_revision), cdate, ctime
+    IF ( TRIM (west_git_revision) /= "unknown" ) THEN
+       WRITE( stdout, '(/5X,"Program ",A," v. ",A," git rev. ",A," starts on ",A9," at ",A9)' ) &
+       & TRIM(code), TRIM(west_version_number), TRIM (west_git_revision), cdate, ctime
     ELSE
        WRITE( stdout, '(/5X,"Program ",A," v. ",A," starts on ",A9," at ",A9)' ) &
        & TRIM(code), TRIM(west_version_number), cdate, ctime
@@ -236,28 +249,51 @@ CONTAINS
     ENDIF 
     !
     IF( mpime == root ) THEN 
-       !
-       CALL json%initialize()
-       !
-       CALL json%add('runjob.startdate', TRIM(cdate) )
-       CALL json%add('runjob.starttime', TRIM(ctime) )
-       CALL json%add('runjob.completed', .FALSE. )
-       CALL json%add('software.package', "WEST" )
-       CALL json%add('software.program', TRIM(code) )
-       CALL json%add('software.version', TRIM(west_version_number) )
-       IF( TRIM (west_svn_revision) /= "unknown" ) CALL json%add('software.westsvn', TRIM(west_svn_revision) )
-       CALL json%add('software.website',"http://www.west-code.org")
-       CALL json%add('software.citation',"M. Govoni et al., J. Chem. Theory Comput. 11, 2680 (2015).")
-       CALL json%add('software.qeversion', TRIM(version_number) )
-       IF( TRIM (svn_revision) /= "unknown" ) CALL json%add('software.qesvn', TRIM(svn_revision) )
-       CALL json%add('config.io.islittleendian', islittleendian() )
-       !
-       OPEN( NEWUNIT=iunit, FILE=TRIM(logfile) )
-       CALL json%print_file( iunit )
-       CLOSE( iunit )
-       !
-       CALL json%destroy()
-       !
+      ! 
+      CALL json%initialize()
+      !
+      CALL json%add('runjob.startdate', TRIM(cdate) )
+      CALL json%add('runjob.starttime', TRIM(ctime) )
+      CALL json%add('runjob.completed', .FALSE. )
+      CALL json%add('software.package', "WEST" )
+      CALL json%add('software.program', TRIM(code) )
+      CALL json%add('software.version', TRIM(west_version_number) )
+      IF( TRIM (west_git_revision) /= "unknown" ) CALL json%add('software.westgit', TRIM(west_git_revision) )
+      CALL json%add('software.website',"http://www.west-code.org")
+      CALL json%add('software.citation',"M. Govoni et al., J. Chem. Theory Comput. 11, 2680 (2015).")
+      CALL json%add('software.qeversion', TRIM(version_number) )
+      IF( TRIM (svn_revision) /= "unknown" ) CALL json%add('software.qesvn', TRIM(svn_revision) )
+      CALL json%add('config.io.islittleendian', islittleendian() )
+      !
+      OPEN( NEWUNIT=iunit, FILE=TRIM(logfile) )
+      CALL json%print_file( iunit )
+      CLOSE( iunit )
+      !
+      CALL json%destroy()
+      !
+      !s = '{ '
+      !s = s // '"startdate" : '   //'"'//TRIM(cdate) //'" , '
+      !s = s // '"startime" : '    //'"'//TRIM(ctime)     //'" , '
+      !s = s // '"package" : '     //'"WEST"'          //' , '
+      !s = s // '"program" : '     //'"'//TRIM(code)      //'" , '
+      !s = s // '"version" : '     //'"'//TRIM(west_version_number) //'" , '
+      !IF( TRIM (west_git_revision) /= "unknown" ) THEN
+      !   s = s // '"git_version" : ' //'"'//TRIM(west_git_revision) //'" , '
+      !ENDIF
+      !s = s // '"website" : '     //'"http://www.west-code.org"' //' , '
+      !s = s // '"citation" : '    //'"M. Govoni et al., J. Chem. Theory Comput. 11, 2680 (2015)."' //' , '
+      !s = s // '"website" : '     //'"http://www.west-code.org"' //' , '
+      !s = s // '"qeversion" : '     //'"'//TRIM(version_number) //'" , '
+      !s = s // '"islittleendian" : ' //ltoa(islittleendian()) //'   '
+      !s = s // '}'
+      !!
+      !IERR = dict_create(attr)
+      !IERR = attr%setitem("type", "intro" )
+      !!
+      !CALL append_log( s, attr )
+      !!
+      !CALL attr%destroy
+      !
     ENDIF 
     !
   END SUBROUTINE 
@@ -274,7 +310,9 @@ CONTAINS
      USE mp_global,        ONLY : nimage,npool,nbgrp,nproc_image,nproc_pool,nproc_bgrp 
      USE mp_world,         ONLY : nproc,mpime,root 
      USE io_push,          ONLY : io_push_title,io_push_bar
-     USE westcom,          ONLY : logfile 
+     USE forpy_mod,        ONLY : dict, dict_create 
+     USE westcom,          ONLY : logfile
+     !USE logfile_mod,      ONLY : append_log, itoa
      !
      IMPLICIT NONE
      !
@@ -285,6 +323,10 @@ CONTAINS
      INTEGER :: nth, ncores 
      TYPE(json_file) :: json
      INTEGER :: iunit
+     !
+     INTEGER :: IERR
+     CHARACTER(LEN=:),ALLOCATABLE :: s
+     TYPE(dict) :: attr
      !
 #if defined(__OPENMP)
      nth = omp_get_max_threads()
@@ -317,7 +359,7 @@ CONTAINS
      CALL io_push_bar()
 #endif
     !
-    IF( mpime == root ) THEN 
+    IF( mpime == root ) THEN
        !
        CALL json%initialize()
        !
@@ -339,6 +381,27 @@ CONTAINS
        !
        CALL json%destroy()
        !
+!       !
+!       s = '{ '
+!       s = s // '"nranks" : '   //ITOA(nproc)      //' , '
+!       s = s // '"nimage" : '   //ITOA(nimage)     //' , '
+!       s = s // '"nimage" : '   //ITOA(nimage)     //' , '
+!       s = s // '"npool" : '    //ITOA(npool)      //' , '
+!       s = s // '"nbgrp" : '    //ITOA(nbgrp)      //' , '
+!       s = s // '"nrg" : '      //ITOA(nproc_bgrp) //' , '
+!#if defined(__OPENMP)
+!       s = s // '"nthreads" : ' //ITOA(nth)        //' , '
+!#endif
+!       s = s // '"nproc" : '    //ITOA(ncores)     //'   '
+!       s = s // '}'
+!       !
+!       IERR = dict_create(attr)
+!       IERR = attr%setitem("type", "parallel" )
+!       !
+!       CALL append_log( s, attr )
+!       !
+!       CALL attr%destroy
+!       !
     ENDIF 
      !
   END SUBROUTINE