+++ Wfreq/wfreq_readin.f90
@@ -35,7 +35,7 @@ SUBROUTINE wfreq_readin()
   !
   ! READ INPUT_WEST
   !
-  CALL fetch_input(1,(/1/),.TRUE.)
+  CALL fetch_input_yml(1,(/1/),.TRUE.,.FALSE.)
   !
   !  read the input file produced by the pwscf program
   !  allocate memory and recalculate what is needed
@@ -51,7 +51,7 @@ SUBROUTINE wfreq_readin()
   !
   ! READ other sections of the input file
   !
-  CALL fetch_input(2,(/2,3/),.TRUE.)
+  CALL fetch_input_yml(2,(/2,3/),.TRUE.,.FALSE.)
   !
   CALL stop_clock('wfreq_readin')
   !